Moved functionality to PartnerCards.js
